add of prev. commit
